requirements.txt
accelerate==1.4.0
black==25.1.0
datasets==3.3.2
fastapi==0.115.12
huggingface-hub==0.29.3
httpx==0.28.1
isort==6.0.1
matplotlib==3.10.1
mlflow==2.20.3
numpy==1.26.4
opencv-python-headless==4.11.0.86
pandas==2.2.3
paddleocr==2.9.1
paddlepaddle==2.6.2
Pillow==11.1.0
pdf2image==1.17.0
pre-commit==4.1.0
psutil==7.0.0
pydantic==2.10.6
PyMuPDF==1.25.4
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytesseract==0.3.13
PyYAML==6.0.2
ruff==0.11.2
seqeval==1.2.2
asgi-lifespan==2.1.0
torch==2.6.0
transformers==4.49.0
uvicorn[standard]==0.34.0
